Add Login Button to Navbar
Initial work on login button, which is a dropdown (owner / employee).  My thought is that there will be a State based in App that will say who is logged in, and as what (Owner / Employee / Manager).  The navbar will be customized based on that property.  But this commit is just a start, we can discuss.