Home Page Idea
This could be a minimal but sufficient home page, a placeholder to refine, or a final solution if we don't have time to improve it.

Normally we should devote several nice pages to selling the website and the business.  But we can do that when we are further along.  Or leave it simple, to keep the grader from having to look at pages with no database element.