Improving showReports
I've separated out the header row to just show props.headers, allowing use of a <thead> section that could allow more formatting options.  Plus it's cleaner.

Im beginning the process of adding Update and Delete to the rows.   The setup will be that if no onDelete or onUpdate function is passed, those buttons will not appear on the rows.