Cleaned up and organized Booking code
Spent a lot of time troubleshooting date issues:

There is currently an issue with dates.  What is currently happening is that SQLDates is adding 1 to day (getDate) and month.  Oddly, this seems to mostly work.  But if you go to Bookings and update a booking such that the endDate = August 1, sqlDate turns it into '2001-07-32'.  I tried to fix it, but ran out of time.  I figured I'd just leave the bug, and we can fix it at our leisure.