Fix json() error
When fetchState.js line 22 was changed from return res to return res.json(), all the front end code that was already doing response.json() started throwing an "json is not a function" error.  Silly, because json() is a function.  I guess javascript doesn't like to parse json twice.  The trick was to remove only those instances using fetchState.  I think I did the right ones, and I tested quite a bit to make sure everything is working.

Also, "Reserve Now" on the homepage was going to /reservations, I switched that to /bookings for now.  I put a TODO to switch it back eventually